se registra y lista paciente, odontologo, y turno

+++ src/main/java/com/backend/clinica/dto/entrada/modificacion/odontologoModificacionEntradaDto.java
@@ -9,7 +9,7 @@ import javax.validation.constraints.NotNull;
 
 public class odontologoModificacionEntradaDto {
     @NotNull
-    private int id;
+    private Long id;
 
     @NotNull
     private String matricula;
@@ -24,18 +24,18 @@ public class odontologoModificacionEntradaDto {
     public odontologoModificacionEntradaDto() {
     }
 
-    public odontologoModificacionEntradaDto(int id, String matricula, String nombre, String apellido) {
+    public odontologoModificacionEntradaDto(Long id, String matricula, String nombre, String apellido) {
         this.id = id;
         this.matricula = matricula;
         this.nombre = nombre;
         this.apellido = apellido;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
